fix the errors i introduced when i tried to fix the merge conflicts
