+++ src/Proxy/IDatabaseService.cs
@@ -7,9 +7,7 @@
     public interface IDatabaseService
     {
         OracleConnection GetConnection();
-    }
-    public interface IDatabaseService
-    {
+   
         int GetIdSequence(string sequenceName);
 
         DataSet ExecuteQuery(string sql);
